Finished shelling out the CPU, just need to fill in the blanks, instruction wise. Also made change from byte to short for all memory... hopefully that will work out ok

+++ src/com/tnes/Constants.java
@@ -118,8 +118,8 @@ public class Constants {
 	/*
 	 * Status Flag
 	 */
-	public static byte PPU_STAT_HIT = (byte) 0x40;
-	public static byte PPU_STAT_VBLANK = (byte) 0x80;
+	public static short PPU_STAT_HIT = (short) 0x40;
+	public static short PPU_STAT_VBLANK = (short) 0x80;
 
 	/*
 	 * Control Register Flags
